# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
      "title": "FastAPI",
      "version": "0.1.0"
    },
    "paths": {
...
      "/payment-conditions/": {
        "get": {
          "tags": [
            "payment-conditions"
          ],
          "summary": "Get Payment Conditions",
          "description": "Get all payment conditions",
          "operationId": "get_payment_conditions_payment_conditions__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/PaymentConditionResponse"
                    },
                    "type": "array",
                    "title": "Response Get Payment Conditions Payment Conditions  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "payment-conditions"
          ],
          "summary": "Create Payment Condition",
          "description": "Create a new payment condition",
          "operationId": "create_payment_condition_payment_conditions__post",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentConditionCreate"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PaymentConditionResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/payment-conditions/{condition_id}": {
        "get": {
          "tags": [
            "payment-conditions"
          ],
          "summary": "Get Payment Condition",
          "description": "Get a specific payment condition by ID",
          "operationId": "get_payment_condition_payment_conditions__condition_id__get",
          "parameters": [
            {
              "name": "condition_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Condition Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PaymentConditionResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "payment-conditions"
          ],
          "summary": "Delete Payment Condition",
          "description": "Delete a payment condition",
          "operationId": "delete_payment_condition_payment_conditions__condition_id__delete",
          "parameters": [
            {
              "name": "condition_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Condition Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "put": {
          "tags": [
            "payment-conditions"
          ],
          "summary": "Update Payment Condition",
          "operationId": "update_payment_condition_payment_conditions__condition_id__put",
          "parameters": [
            {
              "name": "condition_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Condition Id"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentConditionUpdate"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PaymentConditionResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
      "/photos/": {
        "get": {
          "tags": [
            "photos"
          ],
          "summary": "Get Photos",
          "description": "Get all photos",
          "operationId": "get_photos_photos__get",
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "items": {
                      "$ref": "#/components/schemas/PhotoResponse"
                    },
                    "type": "array",
                    "title": "Response Get Photos Photos  Get"
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": [
            "photos"
          ],
          "summary": "Create Photo",
          "description": "Upload a new photo for a report",
          "operationId": "create_photo_photos__post",
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "$ref": "#/components/schemas/Body_create_photo_photos__post"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
      "/photos/{photo_id}": {
        "get": {
          "tags": [
            "photos"
          ],
          "summary": "Get Photo",
          "description": "Get a specific photo by ID",
          "operationId": "get_photo_photos__photo_id__get",
          "parameters": [
            {
              "name": "photo_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Photo Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/PhotoResponse"
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": [
            "photos"
          ],
          "summary": "Delete Photo",
          "description": "Delete a photo by ID",
          "operationId": "delete_photo_photos__photo_id__delete",
          "parameters": [
            {
              "name": "photo_id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "title": "Photo Id"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Successful Response",
              "content": {
                "application/json": {
                  "schema": {
  
                  }
                }
              }
            },
            "422": {
              "description": "Validation Error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/HTTPValidationError"
                  }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "Body_create_photo_photos__post": {
          "properties": {
            "report_id": {
              "type": "integer",
              "title": "Report Id"
            },
            "image": {
              "type": "string",
              "format": "binary",
              "title": "Image"
            }
          },
          "type": "object",
          "required": [
            "report_id",
            "image"
          ],
          "title": "Body_create_photo_photos__post"
        },
...
        "HTTPValidationError": {
          "properties": {
            "detail": {
              "items": {
                "$ref": "#/components/schemas/ValidationError"
              },
              "type": "array",
              "title": "Detail"
            }
          },
          "type": "object",
          "title": "HTTPValidationError"
        },
...
        "PaymentConditionCreate": {
          "properties": {
            "name": {
              "type": "string",
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "text": {
              "type": "string",
              "title": "Text"
            },
            "active": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active",
              "default": true
            }
          },
          "type": "object",
          "required": [
            "name",
            "text"
          ],
          "title": "PaymentConditionCreate"
        },
        "PaymentConditionResponse": {
          "properties": {
            "id": {
              "type": "integer",
              "title": "Id"
            },
            "name": {
              "type": "string",
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "text": {
              "type": "string",
              "title": "Text"
            },
            "active": {
              "type": "boolean",
              "title": "Active"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            }
          },
          "type": "object",
          "required": [
            "id",
            "name",
            "text",
            "active",
            "created_at",
            "updated_at"
          ],
          "title": "PaymentConditionResponse"
        },
        "PaymentConditionUpdate": {
          "properties": {
            "name": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            },
            "description": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Description"
            },
            "text": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Text"
            },
            "active": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Active"
            }
          },
          "type": "object",
          "title": "PaymentConditionUpdate"
        },
        "PhotoResponse": {
          "properties": {
            "path": {
              "type": "string",
              "title": "Path"
            },
            "thumbnail": {
              "type": "string",
              "title": "Thumbnail"
            },
            "report_id": {
              "type": "integer",
              "title": "Report Id"
            },
            "created_at": {
              "type": "string",
              "format": "date-time",
              "title": "Created At"
            },
            "updated_at": {
              "type": "string",
              "format": "date-time",
              "title": "Updated At"
            },
            "id": {
              "type": "integer",
              "title": "Id"
            }
          },
          "type": "object",
          "required": [
            "path",
            "thumbnail",
            "report_id",
            "created_at",
            "updated_at",
            "id"
          ],
          "title": "PhotoResponse",
          "description": "Model for photo response"
        },
...
        "ValidationError": {
          "properties": {
            "loc": {
              "items": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "integer"
                  }
                ]
              },
              "type": "array",
              "title": "Location"
            },
            "msg": {
              "type": "string",
              "title": "Message"
            },
            "type": {
              "type": "string",
              "title": "Error Type"
            }
          },
          "type": "object",
          "required": [
            "loc",
            "msg",
            "type"
          ],
          "title": "ValidationError"
        }
      }
    }
  }